==> Reports/04_plots_tables.rmd <==
---
title: "R Notebook"
output: html_notebook
---

###1. ACF
###2. Rowmean plots 
###3. Pigeon schematic
###4. Groupmeans ( nearest neighbour)
###5. Table S1. 


####1. ACF 

```{R}
load( file.path( PROJHOME , "Results" , "acf-dba.rda" ))
load( file.path( PROJHOME , "Results" , "acf-nn.rda" ))
load( file.path( PROJHOME , "Results" , "acf-ff.rda" ))
load( file.path( PROJHOME , "Results" , "acf-sp.rda" ))
load( file.path( PROJHOME , "Results" , "acf-fb.rda" ))
load( file.path( PROJHOME , "Results" , "acf-lr.rda" ))
load( file.path( PROJHOME , "Results" , "acf-ffu.rda"))
```
 
```{R}


{pdf( file.path( PROJHOME , "Figures" , "FigureS1.pdf"), useDingbats = F)

{  par( mfrow = c(4,2))
  par( mar = c(5,5,1,3))
plot( acf.nn[[1]]$acf, type = "l", lwd = 4, xlab = "Lag", ylab = "Correlation")
abline( v= acf.nn[[2]], col = "red", lwd = 3)

plot( acf.fb[[1]]$acf, type = "l", lwd = 4, xlab = "Lag", ylab = "Correlation")
abline( v= acf.fb[[2]], col = "red", lwd = 3)

plot( acf.lr[[1]]$acf, type = "l", lwd = 4, xlab = "Lag", ylab = "Correlation")
abline( v= acf.lr[[2]], col = "red", lwd = 3)

plot( acf.sp[[1]]$acf, type = "l", lwd = 4, xlab = "Lag", ylab = "Correlation")
abline( v= acf.sp[[2]], col = "red", lwd = 3)

plot( acf.ff[[1]]$acf, type = "l", lwd = 4, xlab = "Lag", ylab = "Correlation")
abline( v= acf.ff[[2]], col = "red", lwd = 3)

plot( acf.dba[[1]]$acf, type = "l", lwd = 4, xlab = "Lag", ylab = "Correlation")
abline( v= acf.dba[[2]], col = "red", lwd = 3)

plot( acf.ffu[[1]]$acf, type = "l", lwd = 4, xlab = "Lag", ylab = "Correlation")
abline( v= acf.ffu[[2]], col = "red", lwd = 3)
}
dev.off()
}
```

####2. Rowmean plots 

```{R}
rm(list=ls())
load( file.path( PROJHOME , "Output" , "GPS-coh.rda"))
load( file.path( PROJHOME , "Output" , "ff-coh.rda"))
load( file.path( PROJHOME , "Output" , "dba-coh.rda"))
load( file.path( PROJHOME , "Output" , "nn-coh.rda")) 
load( file.path( PROJHOME , "Output" , "centroid.rda"))
load( file.path( PROJHOME , "Output" , "dist2cent.rda"))
len = dim(GPS.coh)[1]
centroid = centroid[1:len,]
dist2cent= dist2cent[1:len,]
obj = fb_lr_d2c(GPS.coh,centroid,dist2cent)
fb.coh = abs(obj[[1]])
lr.coh = abs(obj[[2]])
sp.coh = GPS.coh[, "speed",]
```

```{R}


    {par( mfrow = c(3,2))
      par( mar = c(5,5,1.5,3))
      
      time = c(1:nrow(nn.coh))/5/60
      
      plot(rowMeans(nn.coh)~ time , ylab = "Nearest neighbour distance (m)" , xlab = "Time (min)")
      plot(rowMeans(fb.coh)~ time , ylab = "Caudal-cranial \n distance to centroid (m)" , xlab = "Time (min)", ylim = c(0,22))
      plot(rowMeans(lr.coh)~ time , ylab = "Ventral \n distance to centroid (m)" , xlab = "Time (min)", ylim = c(0,22))
      plot(rowMeans(sp.coh)~ time , ylab = "Ground speed (m/s)", xlab = "Time (min)" )
      
      
      time = c(1:nrow(ff.coh))/60
      plot(rowMeans(ff.coh)~ time , ylab = "Flap frequency (Hz)", xlab = "Time (min)")
      plot(rowMeans(dba.coh)~time , ylab = "Dorsal body \n amplitude (mm)", xlab = "Time (min)")

  }
```



###3. Pigeon shematic

#### Load data

```{R}

load( file.path (PROJHOME , "Output" , "nn-coh.rda"))

```

```{R}
floor(floor(dim(nn.coh)[1]/5/60)/3)

mins = 5
step.len = 5*5*60
len = floor(97/mins)
step = rep(step.len, len)
cum.step = c(1,cumsum(step)+1)

mean = rep(NA,len)
sd   = rep(NA,len)
for ( i in 1:len){
  mean[i] = mean(nn.coh[cum.step[i]:cum.step[i+1],])
  sd  [i] = sd  (nn.coh[cum.step[i]:cum.step[i+1],])
}

mean
```

```{R}
if( !file.exists( PROJHOME , "Figures", "pidge.schem.pdf")){
  pdf( file.path (PROJHOME, "Figures" , "pidge.schem.pdf"))
  mean2 = mean/2
  mean2 = mean/2
  {
  par( mar = c(5,5,5,5))
  plot(-15:15,seq(-4,4,length.out = 31) ,type = "n", bty = "n", yaxt = "n", xaxt = "n", ylab = "", xlab = "")
  axis( 2, at= c(-4,-2,0,2), labels = c(0,2,4,6))
  mtext ( "Distance (m)", 2 , line = 3)
  segments( -9,0-mean2[1],-9,0+mean2[1])
  #points(-9,0)
  #segments( -9-sd[1],0 ,-9+sd[1],0)
  
  segments( 9,0-mean2[len],9,0+mean2[len])
  #points(9,0)
  #segments( 9-sd[len],0 ,9+sd[len],0)
  }
  dev.off()
} 
mean2 = mean/2
mean2 = mean/2

par( mar = c(5,5,5,5))
plot(-15:15,seq(-4,4,length.out = 31) ,type = "n", bty = "n", yaxt = "n", xaxt = "n", ylab = "", xlab = "")
axis( 2, at= c(-4,-2,0,2), labels = c(0,2,4,6))
mtext ( "Distance (m)", 2 , line = 3)
segments( -9,0-mean2[1],-9,0+mean2[1])
#points(-9,0)
#segments( -9-sd[1],0 ,-9+sd[1],0)

segments( 9,0-mean2[len],9,0+mean2[len])
#points(9,0)
#segments( 9-sd[len],0 ,9+sd[len],0)

```


###4. Groupmeans ( nearest neighbour, ff, dba)

```{R}




pdf( file.path(PROJHOME , "Figures" , "Group_bin-nn.pdf"), useDingbats = F)
y = bm_plotDS(matr = nn.coh, error = "SE", ylim = c(0,6), x = c(1,97), xlab = "Time (min)", ylab = "Distance to nearest neighbour (m)")
dev.off()
bm_plotDS(matr = nn.coh, error = "SE", ylim = c(0,6), x = c(1,97), xlab = "Time (min)", ylab = "Distance to nearest neighbour (m)")
```

```{R}

pdf( file.path(PROJHOME , "Figures" , "Group_bin_ff.pdf"), useDingbats = F)
y = bm_plotDS(matr = ff.coh, error = "SE", ylim = c(6.5,7.3), x = c(1,97), xlab = "Time (min)", ylab = "Flap frequency (Hz)")
dev.off()
bm_plotDS(matr = ff.coh, error = "SE", ylim = c(6.5,7.3), x = c(1,97), xlab = "Time (min)", ylab = "Flap frequency (Hz)")
```

```{R}


pdf( file.path(PROJHOME , "Figures" , "Group_bin_dba.pdf"), useDingbats = F)
#par(mar = c(3,6,3,11))
y = bm_plotDS(matr = dba.coh, error = "SE", ylim = c(12,15), x = c(1,97), xlab = "Time (min)", ylab = "Dorsal body amplitude (mm)", pch = )
dev.off()
par(mar = c(3,3,3,11))
bm_plotDS(matr = dba.coh, error = "SE", ylim = c(12,15), x = c(1,97), xlab = "Time (min)", ylab = "Dorsal body amplitude (mm)")

```

###5. Table S1 

```{R}

load( file.path ( PROJHOME , "Results" , "stats-nn-sca.rda"))
load( file.path ( PROJHOME , "Results" , "stats-fb-sca.rda"))
load( file.path ( PROJHOME , "Results" , "stats-lr-sca.rda"))
load( file.path ( PROJHOME , "Results" , "stats-sp-sca.rda"))
load( file.path ( PROJHOME , "Results" , "stats-ff-sca.rda"))
load( file.path ( PROJHOME , "Results" , "stats-dba-sca.rda"))
load( file.path ( PROJHOME , "Results" , "stats-ffu-sca.rda"))



load( file.path ( PROJHOME , "Results" , "stats-nn-plot.rda"))
load( file.path ( PROJHOME , "Results" , "stats-fb-plot.rda"))
load( file.path ( PROJHOME , "Results" , "stats-lr-plot.rda"))
load( file.path ( PROJHOME , "Results" , "stats-sp-plot.rda"))
load( file.path ( PROJHOME , "Results" , "stats-ff-plot.rda"))
load( file.path ( PROJHOME , "Results" , "stats-dba-plot.rda"))
load( file.path ( PROJHOME , "Results" , "stats-ffu-plot.rda"))


```

```{R}

med.mea = function( x,dp =3){

  if(! any(is.na(x))){
return(ifelse( shapiro.test(x)$p.value < 0.05, round( median( x ),dp), round( mean (x), dp)))
  } else {
    return(NA)
  }
}
```



```{R}

dim.names = c( "CIs", "fixed.ef", "rand.ef", "t","r2","AIC")

results.mat = matrix( NA, 7,10)

  results.mat[1,1] = med.mea(stats.nn.sca[[5]][,1])
  results.mat[1,2] = med.mea(stats.nn.sca[[5]][,2])
  results.mat[1,3] = med.mea(stats.nn.sca[[1]][,1])
  results.mat[1,4] = med.mea(stats.nn.sca[[1]][,2])
  results.mat[1,5] = med.mea(stats.nn.plot[[5]][,1])
  results.mat[1,6] = med.mea(stats.nn.plot[[5]][,2])
  results.mat[1,7] = med.mea(stats.nn.plot[[1]][,1])
  results.mat[1,8] = med.mea(stats.nn.plot[[1]][,2])
  
  results.mat[2,1] = med.mea(stats.fb.sca[[5]][,1])
  results.mat[2,2] = med.mea(stats.fb.sca[[5]][,2])
  results.mat[2,3] = med.mea(stats.fb.sca[[1]][,1])
  results.mat[2,4] = med.mea(stats.fb.sca[[1]][,2])
  results.mat[2,5] = med.mea(stats.fb.plot[[5]][,1])
  results.mat[2,6] = med.mea(stats.fb.plot[[5]][,2])
  results.mat[2,7] = med.mea(stats.fb.plot[[1]][,1] ) 
  results.mat[2,8] = med.mea(stats.fb.plot[[1]][,2])
  
  results.mat[3,1] = med.mea(stats.lr.sca[[5]][,1])
  results.mat[3,2] = med.mea(stats.lr.sca[[5]][,2])
  results.mat[3,3] = med.mea(stats.lr.sca[[1]][,1])
  results.mat[3,4] = med.mea(stats.lr.sca[[1]][,2])
  results.mat[3,5] = med.mea(stats.lr.plot[[5]][,1])
  results.mat[3,6] = med.mea(stats.lr.plot[[5]][,2])
  results.mat[3,7] = med.mea(stats.lr.plot[[1]][,1] ) 
  results.mat[3,8] = med.mea(stats.lr.plot[[1]][,2])
  
  results.mat[4,1] = med.mea(stats.sp.sca[[5]][,1])
  results.mat[4,2] = med.mea(stats.sp.sca[[5]][,2])
  results.mat[4,3] = med.mea(stats.sp.sca[[1]][,1])
  results.mat[4,4] = med.mea(stats.sp.sca[[1]][,2])
  results.mat[4,5] = med.mea(stats.sp.plot[[5]][,1])
  results.mat[4,6] = med.mea(stats.sp.plot[[5]][,2])
  results.mat[4,7] = med.mea(stats.sp.plot[[1]][,1])
  results.mat[4,8] = med.mea(stats.sp.plot[[1]][,2])
  results.mat[4,9] = med.mea(stats.sp.sca[[7]],dp=3)
  results.mat[4,10]= med.mea(stats.sp.plot[[7]])
  
  results.mat[5,1] = med.mea(stats.ff.sca[[5]][,1])
  results.mat[5,2] = med.mea(stats.ff.sca[[5]][,2])
  results.mat[5,3] = med.mea(stats.ff.sca[[1]][,1])
  results.mat[5,4] = med.mea(stats.ff.sca[[1]][,2])
  results.mat[5,5] = med.mea(stats.ff.plot[[5]][,1])
  results.mat[5,6] = med.mea(stats.ff.plot[[5]][,2])
  results.mat[5,7] = med.mea(stats.ff.plot[[1]][,1])
  results.mat[5,8] = med.mea(stats.ff.plot[[1]][,2])
  results.mat[5,9] = med.mea(stats.ff.sca[[7]],dp=3)
  results.mat[5,10]= med.mea(stats.ff.plot[[7]],dp=3)
  
  
  results.mat[6,1] = med.mea(stats.dba.sca[[5]][,1])
  results.mat[6,2] = med.mea(stats.dba.sca[[5]][,2])
  results.mat[6,3] = med.mea(stats.dba.sca[[1]][,1])
  results.mat[6,4] = med.mea(stats.dba.sca[[1]][,2])
  results.mat[6,5] = med.mea(stats.dba.plot[[5]][,1])
  results.mat[6,6] = med.mea(stats.dba.plot[[5]][,2])
  results.mat[6,7] = med.mea(stats.dba.plot[[1]][,1])
  results.mat[6,8] = med.mea(stats.dba.plot[[1]][,2])
  
  results.mat[7,1] = med.mea(stats.ffu.sca[[5]][,1])
  results.mat[7,2] = med.mea(stats.ffu.sca[[5]][,2])
  results.mat[7,3] = med.mea(stats.ffu.sca[[1]][,1])
  results.mat[7,4] = med.mea(stats.ffu.sca[[1]][,2])
  results.mat[7,5] = med.mea(stats.ffu.plot[[5]][,1])
  results.mat[7,6] = med.mea(stats.ffu.plot[[5]][,2])
  results.mat[7,7] = med.mea(stats.ffu.plot[[1]][,1])
  results.mat[7,8] = med.mea(stats.ffu.plot[[1]][,2])

  write.csv( results.mat, file = file.path(PROJHOME, "Results", "results-mat.csv"))

```

```{R}



```
