==> Reports/04_stats.rmd <==
---
title: "Statistical analysis "
output: html_notebook
---



####1. Nearest neighbour
#####a. Group level root function
#####b. individual level models
####2. Flap frequency
####3. Group speed 
####4. Flap frequency for solo flight (after split)
####5. Dist to the centroid
#####a. Frontback
#####b. Leftright
####6. Dorsal body amplitude

```{R, echo = F}

# Clear working environment
rm(list=ls())


# objects 



# libraries 

library(lme4)
library(sjmisc)
library(sjPlot)
library(plyr)
library(MuMIn)
library(longFlight)
library(mgcv)

#data

load( file.path( PROJHOME , "Output" , "GPS-coh.rda"))
load( file.path( PROJHOME , "Output" , "ff-coh.rda"))
load( file.path( PROJHOME , "Output" , "dba-coh.rda"))
load( file.path( PROJHOME , "Output" , "nn-coh.rda")) 

coh = dim(nn.coh)[1] # time of split
len = ncol(nn.coh)+1 # number of individuals plus a group level measure. 
cols = c(rainbow(6),"black")
time = c(1:coh)/60/5


```

### 1. nearest neighbour 

#### a) Group level root function


```{R}

# group mean distance
y = (rowMeans(nn.coh))

# root transforms to try
roots = seq(1,1.4,0.1)

# object for r squared 
r.sq = rep(NA,length(roots))

# loop to compare root transforms 

for (  i in 1:length(roots)){
  
  x = c(1:length(y))/5/60
  mod2 = lm(y^(1/roots[i])~x)
  
  r.sq[i] = (summary(mod2)$r.squared)
  plot(y~x)
  
  y2 = mod2$coefficients[1] + mod2$coefficients[2]*  x
  lines(y2^roots[i]~x,col = "red")
}
```

```{R}
roots[which.max(r.sq)]

```

```{R}
mod2 = lm( lm(y^(1/ roots[which.max(r.sq)])~x) )

summary(mod2)
```

#### Plot 

```{R}

pdf( file.path(PROJHOME , "Figures" , "Group_dist2neighbour.pdf"))
y = bm_plotDS(matr = nn.coh, error = "SD", ylim = c(0,10), x = c(1,97), xlab = "Time (min)", ylab = "Distance to nearest neighbour (m)")
dev.off()
bm_plotDS(matr = nn.coh, error = "SD", ylim = c(0,10), x = c(1,97), xlab = "Time (min)", ylab = "Distance to nearest neighbour (m)")
```

#### b) Individual level models 

#### Autocorrelation 

```{R}

acf(as.vector(nn.coh), lag.max = 1000)

```


Seems to drop around 300 (to be safe, 400)

# Which model is best? 

```{R}

AICs = matrix(NA,100,3)


for ( i in 1:100){
  mod= try( model2(nn.coh,400,scale = T))
  AICs[i,] = mod[[3]]
}

```

```{R}

whi.mi = which.min(colMeans(AICs))
print( c("simple linear model is best", "random intercepts only","random intercepts and slopes")[whi.mi])
```

```{R}

CIs = matrix( NA, 100, 2)
for ( i in 1:100){
  mod = model2( nn.coh, 400, scale = T, specify.model = whi.mi)
  CIs[i,] = mod[[2]]
}

colMeans(CIs)
```

```{R}

CIs.method2 = matrix( NA, 100, 2)
for ( i in 1:100){
  mod = model2( nn.coh, 400, scale = T, specify.model = whi.mi, conf.method = "confint")
  CIs.method2[i,] = mod[[2]]
}

colMeans(CIs.method2)

```

```{R}

t.test( CIs[,1],CIs.method2[,1])
t.test( CIs[,2],CIs.method2[,2])

```

#### Build dataframe 


```{R, echo = F}

# Sample length (sam.len)

sam.len = nrow ( nn.coh)/400
sam = sample( 1:nrow(nn.coh), sam.len)
nn = data.frame ( nn = as.vector(nn.coh[sam,]), 
                  id = as.factor(rep (1: ncol(nn.coh), each = sam.len)),
                  time = rep(time[sam], ncol(nn.coh)))

#head(nn)
```

#### plot

```{R, echo = F}

# fit glmer

m2 =lmer( data = nn, nn ~ time + (1+ time|id)  ) 

a = fixef(m2)
b = ranef(m2, condVar = T)

# Extract the variances of the random effects
qq <- attr(b[[1]], "postVar")
e=(sqrt(qq)) 
e=e[2,2,] #here we want to access the Petal.Weigth, which is stored in column 2 in b[[1]], that's why I use the [,2,2]
#calculate CI's
liminf=(b[[1]][2]+a[2])-(e*2)
mean_=(b[[1]][2]+a[2])
limsup=(b[[1]][2]+a[2])+(e*2)
#Plot betas and its errors

```

```{R}


#make final plot
plot(nn$nn ~ nn$time, col = cols, las = 1, ylim = c(0,6), xlab = "Time (min)", 
     type = "n",
     ylab = "Nearest neighbour distance (m)")
#and plot each random slope

for ( i in 1:6){
  
  abline(a = as.numeric(b[[1]][i,1])+ as.numeric(a[1]), b= mean_$time[i], col = cols[i], lwd=2 )
  
}
#and general response
abline(a, lwd = 4)

pdf( file.path ( PROJHOME , "Figures" , "nn.mixed.plot.pdf"))
plot(nn$nn ~ nn$time, col = cols, las = 1, ylim = c(0,6), xlab = "Time (min)", 
     type = "n",
     ylab = "Nearest neighbour distance (m)")
for ( i in 1:6){
  abline(a = as.numeric(b[[1]][i,1])+ as.numeric(a[1]), b= mean_$time[i], col = cols[i], lwd=2 )
}
abline(a, lwd = 4)
dev.off()
```


###2. Flap frequency 


#### autocorrelation 

```{R, echo = F}
acf( as.vector(ff.coh), lag.max =  200, ylim = c(0,0.4))
```

50 seems a conservative estimate 

# Which model is best? 

```{R}

AICs = matrix(NA,100,3)


for ( i in 1:100){
  mod= try( model2(ff.coh,50,scale = T))
  AICs[i,] = mod[[3]]
}

```

```{R}

whi.mi = which.min(colMeans(AICs))
print( c("simple linear model is best", "random intercepts only","random intercepts and slopes")[whi.mi])
```

```{R}

CIs = matrix( NA, 100, 2)
for ( i in 1:100){
  mod = model2( ff.coh, 50, scale = T, specify.model = whi.mi)
  CIs[i,] = mod[[2]]
}

colMeans(CIs)
```

```{R}

CIs.method2 = matrix( NA, 100, 2)
for ( i in 1:100){
  mod = model2( ff.coh, 50, scale = T, specify.model = whi.mi, conf.method = "confint")
  CIs.method2[i,] = mod[[2]]
}

colMeans(CIs.method2)

```

```{R}

t.test( CIs[,1],CIs.method2[,1])
t.test( CIs[,2],CIs.method2[,2])

```

#### Build dataframe (part 2)

```{R, echo = F} 

# Sample length (sam.len)

ind.len = nrow ( ff.coh)
sam.len = ind.len/50
sam = sample( 1:ind.len, sam.len)

ff = data.frame( ff = as.vector(ff.coh[sam,]),
                 id = rep(1:6, each = length(sam)),
                 time = rep(sam,6)/60)

```


#### Plot

```{R, echo = F}

fit <- lmer(ff ~ time + (1|id), data = ff)

a = fixef(fit)
b = ranef(fit, condVar = T)

#make final plot
{plot(ff$ff ~ ff$time, col = as.numeric(ff$id)+1, las = 1, ylim = c(6.6,7.3), xlab = "Time (min)", 
      ylab = "Flap frequency (Hz)", type = "n")
  #and plot each random slope
  for ( i in 1:6){
    abline(a = as.numeric(b[[1]][i,1])+ as.numeric(a[1]), b= a[2], col = cols[i], lwd=2 )
  }
}
#and general response
abline(a, lwd = 4)

```

```{R}
if( ! file.exists( file.path(PROJHOME , "Figures" , "ff.pdf"))){
  pdf( file.path ( PROJHOME , "Figures", "ff.pdf"))
  {plot(ff$ff ~ ff$time, col = as.numeric(ff$id)+1, las = 1, ylim = c(6.6,7.3), xlab = "Time (min)", 
      ylab = "Flap frequency (Hz)", type = "n")
  for ( i in 1:6){
    abline(a = as.numeric(b[[1]][i,1])+ as.numeric(a[1]), b= a[2], col = cols[i], lwd=2 )
  }
}
abline(a, lwd = 4)
dev.off()
}
```

###3. Group speed 


#### Autocorrelation

```{R, echo = F}
acf( as.vector(GPS.coh[, "speed",] ), lag.max = 400)

```

250 seems fine here 

# Which model is best? 

```{R}

sp.coh = GPS.coh[, "speed",]

AICs = matrix(NA,100,3)


for ( i in 1:100){
  mod= try( model2(sp.coh,250,scale = T))
  AICs[i,] = mod[[3]]
}

```

```{R}

whi.mi = which.min(colMeans(AICs))
print( c("simple linear model is best", "random intercepts only","random intercepts and slopes")[whi.mi])
```

```{R}

CIs = matrix( NA, 100, 2)
for ( i in 1:100){
  mod = model2( sp.coh, 250, scale = T, specify.model = whi.mi)
  CIs[i,] = mod[[2]]
}

colMeans(CIs)
```




```{R, echo = F}

sam.len = dim(sp.coh)[1]/250
sam = sample( 1:dim(sp.coh)[1], sam.len)


sp = data.frame( sp =  as.vector(sp.coh[sam,] ),
                 time = rep(sam, 6)/5/60,
                 id =  as.factor(rep(1:6, each = length(sam))))

```

```{R, echo = F}

fit = lm(sp$sp~sp$time)

if( ! file.exists( file.path ( PROJHOME , "Figures" , "Groupspeed.pdf"))){
  
  pdf( file.path (PROJHOME , "Figures", "Groupspeed.pdf"))
  plot(sp$sp~sp$time, 
       xlab = "Time (min)", 
       ylab = "Ground speed (m/s)"
       # type = "n" ,
       # ylim = c(10,20)
  )
  abline( fit$coefficients[1],fit$coefficients[2], lwd = 2)
  dev.off()
}
```



### 5. Dist to the centroid


```{R}

load( file.path( PROJHOME , "Output" , "centroid.rda"))
load( file.path( PROJHOME , "Output" , "dist2cent.rda"))


len = dim(GPS.coh)[1]

centroid = centroid[1:len,]
dist2cent= dist2cent[1:len,]

obj = fb_lr_d2c(GPS.coh,centroid,dist2cent)

fb.coh = abs(obj[[1]])
lr.coh = abs(obj[[2]])

```


####a. Front/back

#### Autocorrelation 


```{R, echo = F}

acf(na.omit(as.vector(abs(frontback.res.cent) )), lag.max = 1000)

```

600

# Which model is best? 

```{R}

AICs = matrix(NA,100,3)


for ( i in 1:100){
  mod= try( model2(fb.coh,600,scale = T))
  AICs[i,] = mod[[3]]
}

```

```{R}

whi.mi = which.min(colMeans(AICs))
print( c("simple linear model is best", "random intercepts only","random intercepts and slopes")[whi.mi])
```

```{R}

CIs = matrix( NA, 100, 2)
for ( i in 1:100){
  mod = model2(fb.coh, 600, scale = T, specify.model = whi.mi)
  CIs[i,] = mod[[2]]
}

colMeans(CIs)
```



####b. left/right 

```{R}
acf(na.omit(as.vector(lr.coh )), lag.max = 1000)

```
200

# Which model is best? 

```{R}

AICs = matrix(NA,100,3)


for ( i in 1:100){
  mod= try( model2(lr.coh,200,scale = T))
  AICs[i,] = mod[[3]]
}

```

```{R}

whi.mi = which.min(colMeans(AICs))
print( c("simple linear model is best", "random intercepts only","random intercepts and slopes")[whi.mi])
```

```{R}

CIs = matrix( NA, 100, 2)
for ( i in 1:100){
  mod = model2(lr.coh, 200, scale = T, specify.model = whi.mi)
  CIs[i,] = mod[[2]]
}

colMeans(CIs)
```


```{R}

r.squaredGLMM( mod.fb[[3]])

```




###6. Dorsal body amplitude

#### Autocorrelation 

```{R}

acf( na.omit(as.vector(dba.coh)), lag.max = 1000)

```

100 seems fine

# Which model is best? 

```{R}

AICs = matrix(NA,100,3)


for ( i in 1:100){
  mod= try( model2(dba.coh,100,scale = T))
  AICs[i,] = mod[[3]]
}

```

```{R}

whi.mi = which.min(colMeans(AICs))
print( c("simple linear model is best", "random intercepts only","random intercepts and slopes")[whi.mi])
```

```{R}

CIs = matrix( NA, 100, 2)
for ( i in 1:100){
  mod = model2( dba.coh, 100, scale = T, specify.model = whi.mi)
  CIs[i,] = mod[[2]]
}

colMeans(CIs)
```

```{R}

if ( !file.exists( file.path(PROJHOME , "Figures", "dba.pdf"))){
  pdf( file.path(PROJHOME , "Figures", "dba.pdf"))
  bm_plotDS(matr = dba.coh , return_means = F, total.bins = 98, error = "SD", ylim= c(9,20),ylab = "Dorsal body amplitude (mm)", xlab = "Time (min)")
  dev.off()
} else {
  pdf( file.path(PROJHOME , "Figures", "dba.pdf"))
  bm_plotDS(matr = dba.coh , return_means = F, total.bins = 98, error = "SD", ylim= c(9,20),ylab = "Dorsal body amplitude (mm)", xlab = "Time (min)")
}

```
