==> Reports/06_Mass.rmd <==


# MASS

```{R}

### Objects 
load( file.path( PROJHOME, "Output", "gGPS", "g2.34.rda")) # load group flight
pigeon = names(data[1,1,]) # pigeon names
cols = c(rainbow(6),"black")

### DATA
load( file.path (PROJHOME  , "Data" , "mass.rda" ))

data = mass[ mass$Pigeon == pigeon[1] |
               mass$Pigeon == pigeon[2] |
               mass$Pigeon == pigeon[3] |
               mass$Pigeon == pigeon[4] |
               mass$Pigeon == pigeon[5] |
               mass$Pigeon == pigeon[6],]

## LIBRARIES

library(stringr)


```

```{R}


dates = str_split_fixed(data$Date , "\\." , 2)

data$Day = ifelse ( dates[,2] == "7" , as.numeric(dates[,1]) +30,
                    ifelse(dates[,2] == "8", as.numeric(dates[,1]) +61,
                           ifelse(dates[,2] == "9", as.numeric(dates[,1]) +92,
                                  as.numeric(dates[,1]) )))

cols = rainbow(6)

mean.mass= rep(NA, length(unique(data$Pigeon)))

# time intervals (for number of weighings and interval bet

ti = list()

{plot(data$Weight ~  data$Day , xlim = c(40,57) ,type = "n", ylab = "Mass (g)", xlab = "Days since June 1st 2017")

  for ( i in 1:length(unique(data$Pigeon))){

    lines(data$Weight[data$Pigeon== unique(data$Pigeon)[i]] ~ data$Day[data$Pigeon== unique(data$Pigeon)[i]], col = cols[i])
    mean.mass[i] = mean(data$Weight[data$Pigeon== unique(data$Pigeon)[i]])
    ti[[i]] = data$Day[data$Pigeon== unique(data$Pigeon)[i]]
  }
  abline (v =48, lty = 2)
    text   (62, 520 , labels = "Long ass flight (100km)", cex = 0.8)
}

```

```{R}



mass = data.frame( before =  data$Weight[data$Day == 41][order(as.numeric(data$Pigeon[data$Day == 41]))],
                   after  =  data$Weight[data$Day == 50][order(as.numeric(data$Pigeon[data$Day == 50]))], 
                   after2 =  data$Weight[data$Day == 56][order(as.numeric(data$Pigeon[data$Day == 56]))])

mass$diff =  mass$after - mass$before

mass$diff

plot ( c(1,2), c(mass$before[1],mass$after[1]), type = "l", ylim = c(380,480), col = cols[1], ylab = "Mass (g)", xlab = "", xaxt = "n" , xlim = c(0.8,2.2))
axis(1 , at = c(1,2) , labels = c("before" , "after"))
for( i in 2:6){
  lines(c(1,2), c(mass$before[i],mass$after[i]), type = "l", col = cols[i], lwd = 2)
}


pdf ( file.path(PROJHOME  , "Figures", "mass.pdf"))
plot ( c(1,2), c(mass$before[1],mass$after[1]), type = "l", ylim = c(380,480), col = cols[1], ylab = "Mass (g)", xlab = "", xaxt = "n" , xlim = c(0.8,2.2))
axis(1 , at = c(1,2) , labels = c("before" , "after"))
for( i in 2:6){
  lines(c(1,2), c(mass$before[i],mass$after[i]), type = "l", col = cols[i], lwd = 2)
}
dev.off()

#plot( data$Pigeon)
```

```{R}

pdf( file.path(PROJHOME, "Figures", "mass.pdf"))
{plot ( c(11,20,26), c(mass$before[1],mass$after[1],mass$after2[1]), ylim = c(380,480), col = cols[1], ylab = "Mass (g)", xlab = "", xaxt = "n" ,type = "n", xlim = c(10,27))
axis(1 , at = c(11,20,26) ,labels = c("11th July" , "20th July" , "26th July"), main = i)
set.seed(18)
for( i in sample(1:6)){
 points(jitter(c(11,20,26),0.14), c(mass$before[i],mass$after[i],mass$after2[i]), col = cols[i], lwd = 2, pch= 19, cex = 2)
}
abline ( v = 18, lty = 2)
}
dev.off()

```

```{R}

data.frame =  c( mass$before, mass$after )

t.test( mass$before , mass$after, paired = T)

```

```{R}
boxplot(rep(NA,6) ,mass$before, mass$after, rep(NA,6), xaxt = "n")
axis(1, at = c(2,3) , labels = c("before", "after"))
```

```{R}

{

  par(mfrow = c(2,1))

  {plot(data$Weight ~  data$Day , type = "n", ylab = "Mass (g)", xlab = "Days since June 1st 2017", main = "Group 1")

    for ( i in g1){

      points(data$Weight[data$Pigeon== unique(data$Pigeon)[i]] ~ data$Day[data$Pigeon== unique(data$Pigeon)[i]], col = cols[i])
      mean.mass[i] = mean(data$Weight[data$Pigeon== unique(data$Pigeon)[i]])
      ti[[i]] = data$Day[data$Pigeon== unique(data$Pigeon)[i]]
    }

    abline (v =51, lty = 2)
    text   (62, 520 , labels = "Long ass flight (100km)", cex = 0.8)
  }

  {plot(data$Weight ~  data$Day , type = "n", ylab = "Mass (g)", xlab = "Days since June 1st 2017", main = "Group 2")

    for ( i in g2){

      points(data$Weight[data$Pigeon== unique(data$Pigeon)[i]] ~ data$Day[data$Pigeon== unique(data$Pigeon)[i]], col = cols[i])
      mean.mass[i] = mean(data$Weight[data$Pigeon== unique(data$Pigeon)[i]])
      ti[[i]] = data$Day[data$Pigeon== unique(data$Pigeon)[i]]
    }

    abline (v = 48, lty = 2)
    text   (62, 520 , labels = "Long ass flight (200km)", cex = 0.8)
  }


}

```



```{R}

load( file.path( PROJHOME , "Output" , "gGPS" , "g2.34.rda") )

dim(data)
data[1:20,,1]
dis.array = array( NA, c(23401, 7,7))

for ( i in 1:7){
  for ( j in 1:7){
    if ( i != j)
      dis.array[,j,i] = get_dist(data[,1,i], data[,2,i], data[,1,j], data[,2,j], method = "distance")
  }
  print(i)
}

mat = matrix(NA, 23401, 7)
mat2 = matrix(NA, 23401, 7)
for ( i in 1:7){
  mat[,i] = apply( dis.array[,-i,i], 1, min)
  mat2[,i]= apply( dis.array[,-i,i], 1, mean)
}

```

```{R}
{
  par(mfrow = c(4,2))
  for ( i in 1:7){
    bm_plotDS(mat[,i], 50, plot.error = F, ylab = "Closest neighbour (m)")
  }
}
```


```{R}

{
  par(mfrow = c(4,2))
  for ( i in 1:7){
    bm_plotDS(mat2[,i], 50, plot.error = F, ylab = "man dist to all neighbours(m)")
  }
}

```



```{R}


```
